+++ packages/@poupe-vue/src/components/icon.vue
@@ -11,7 +11,7 @@ export interface IconProps extends Omit<IconifyIconProps, 'icon'> {
 <script setup lang="ts">
 import { Icon as IconifyIcon } from '@iconify/vue';
 import { computed } from 'vue';
-import { usePoupeIcons } from '@/composables/use-icons';
+import { usePoupeIcons } from '../composables/use-icons';
 
 const props = defineProps<IconProps>();